Split PlayerStandingsCalculator into two classes
Split class into one class that fetches advancing players and one that fetches ordered list of player standings in a certain group.
