+++ Slask.Domain/Groups/GroupUtility/PlayerStandingsSolver.cs
@@ -1,0 +1,51 @@
+﻿using Slask.Domain.Groups.Bases;
+using Slask.Domain.Rounds.Bases;
+using Slask.Domain.Utilities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Slask.Domain.Groups.GroupUtility
+{
+    // CREATE TESTS
+    public static class PlayerStandingsSolver
+    {
+        public static List<PlayerStandingEntry> FetchFrom(GroupBase group)
+        {
+            List<PlayerStandingEntry> playerStandings = CreateStandingsList(group);
+
+            AggregateWinsForPlayerStandingEntries(playerStandings, group);
+
+            return playerStandings.OrderByDescending(player => player.Wins).ToList();
+        }
+
+        private static List<PlayerStandingEntry> CreateStandingsList(GroupBase group)
+        {
+            List<PlayerStandingEntry> playerStandings = new List<PlayerStandingEntry>();
+
+            foreach (PlayerReference participant in group.PlayerReferences)
+            {
+                playerStandings.Add(PlayerStandingEntry.Create(participant));
+            }
+
+            return playerStandings;
+        }
+
+        private static void AggregateWinsForPlayerStandingEntries(List<PlayerStandingEntry> playerStandings, GroupBase group)
+        {
+            foreach (Match match in group.Matches)
+            {
+                PlayerReference winner = match.GetWinningPlayer().PlayerReference;
+                PlayerStandingEntry playerStandingEntry = playerStandings.Find(player => player.PlayerReference.Name == winner.Name);
+
+                if (playerStandingEntry == null)
+                {
+                    // LOG Error: Failed to find player reference when calculating player standings for some reason
+                    throw new Exception("Failed to find player reference when calculating player standings for some reason");
+                }
+
+                playerStandingEntry.AddWin();
+            }
+        }
+    }
+}
